Extract expandable action element into reusable component
